// src/app/modules/academics/attendance/model/student-leave-request-search-view.ts
import { PagingParams } from 'app/models/paging-params';
import { SortOrder } from 'app/models/sort-order';

/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**StudentLeaveRequestSeachView */
export interface StudentLeaveRequestSeachView extends PagingParams {
    /**Gets or Sets StudentNames */
    studentNames?: string[];
    /**Gets or Sets ClassIds */
    ClassIds?: string[];
    /**Gets or Sets SectionIds */
    sectionIds?: string[];
    /**Gets or Sets AdmissionNumbers */
    admissionNumbers?: string[];
    /**Gets or Sets FromDate for fromDateBegin */
    fromDateBegin?: Date;
    /**Gets or Sets ToDate for fromDateEnd */
    fromDateEnd?: Date;
    /**Gets or Sets FromDate for toDateBegin */
    toDateBegin?: Date;
    /**Gets or Sets ToDate for toDateEnd */
    toDateEnd?: Date;
    /**Gets or Sets Reasons */
    reasons?: string[];
    /**Gets or Sets StatusIds */
    statusIds?: string[];
    /**Gets or Sets SortColumn */
    sortOrder: 0 | 1 | -1;
    /**Gets or Sets SortBy-"StudentName ParentName AdmissionNumber Class Status DOB" */
    sortBy: any;
}